+++ lib/yt/response.rb
@@ -40,6 +40,10 @@ module Yt
       default_params(options).merge playlist_id: options[:playlist_id]
     end
 
+    def video_threads_params(options)
+      default_params(options).merge video_id: options[:video_id]
+    end
+
     def resource_params(options)
       default_params(options).merge id: options[:ids].join(',')
     end