+++ spec/support/server.rb
@@ -22,6 +22,9 @@ RSpec.configure do |config|
   $existing_item_id      = 'UEwtTGVUdXRjOUdSS0QzeUJEaG5SRl95RThVVGFRSTVKZi4yODlGNEE0NkRGMEEzMEQy'
   $unknown_item_id       = 'invalid-id-'
 
+  $existing_thread_id    =  'z121srzx5oqiyrbce23nevcwrpqfenix004'
+  $unknown_thread_id     =  'z121srzx5oqiyrbce23nevcwrpqfe'
+
   config.before(:example, :requests) do |example|
     count = example.metadata[:requests]
     expect(Net::HTTP).to receive(:start).exactly(count).times.and_call_original