+++ CHANGELOG.md
@@ -6,6 +6,11 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## 0.1.3 - 2017-05-23
+
+* [FEATURE] Add Comment and CommentThread
+* [FEATURE] Add Channel#featured_channels_title and Channel#featured_channels_urls
+
 ## 0.1.2 - 2017-04-06
 
 * [BUGFIX] Fix cases like `channel.select(:snippet).view_count` where attribute does not belong to any selected part.